Rewrite label names to be Prometheus compliant
I am running into issues where my OpenTSDB tags with periods in their names are failing to be scraped by Prometheus. It seems like the most logical place to fix that is in this exporter. This approach takes any tag name that would fail validation in Prometheus and rewrites it to pass.